requirements/vm.txt
Django==1.10
requests==2.9.1
dj_database_url
beautifulsoup4==4.4.1
djangorestframework
markdown
django-filter